extract out getAppMenu
